fix: handle UInt8 overflow in TypedStreamParser for long messages
Fix crash when parsing attributedBody BLOBs longer than 256 bytes.
The parser was attempting to cast segment.count - 1 to UInt8 without
checking bounds, causing a fatal error for long messages.

Instead of adding a bounds check, convert segment[0] to Int for
comparison. This is more general (works for any size segment), clearer
(comparing like types), and maintains the original logic correctly.

Add test case for long repeated pattern messages that reproduces the bug.

+++ Sources/IMsgCore/TypedStreamParser.swift
@@ -14,7 +14,8 @@ enum TypedStreamParser {
         let sliceStart = index + 2
         if let sliceEnd = findSequence(end, in: bytes, from: sliceStart) {
           var segment = Array(bytes[sliceStart..<sliceEnd])
-          if segment.count > 1, segment[0] == UInt8(segment.count - 1) {
+          // Check if first byte equals length prefix (convert byte to Int for comparison)
+          if segment.count > 1, Int(segment[0]) == segment.count - 1 {
             segment.removeFirst()
           }
           let candidate = String(decoding: segment, as: UTF8.self)
